DreiDLAm: fixed
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@5974 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/host/Rapidshare.java
@@ -1039,7 +1039,7 @@ public class Rapidshare extends PluginForHost {
     }
 
     // @Override
-    public AccountInfo fetchAccountInfo(Account account) throws Exception {        
+    public AccountInfo fetchAccountInfo(Account account) throws Exception {
         AccountInfo ai = new AccountInfo(this, account);
         String api = "http://api.rapidshare.com/cgi-bin/rsapi.cgi?sub=getaccountdetails_v1&login=" + Encoding.urlEncode(account.getUser()) + "&password=" + Encoding.urlEncode(account.getPass()) + "&type=prem";
         br.getPage(api);
@@ -1052,13 +1052,13 @@ public class Rapidshare extends PluginForHost {
         String[][] matches = br.getRegex("(\\w+)=([^\r^\n]+)").getMatches();
         HashMap<String, String> data = getMap(matches);
 
-        ai.setTrafficLeft((Long.parseLong(data.get("premkbleft")) / 1000) * 1024l * 1024l);
+        ai.setTrafficLeft((long)(Long.parseLong(data.get("premkbleft"))/1000.0) * 1024l*1024l);
         ai.setTrafficMax(25 * 1024 * 1024 * 1024l);
         ai.setFilesNum(Long.parseLong(data.get("curfiles")));
         ai.setPremiumPoints(Long.parseLong(data.get("points")));
         ai.setNewPremiumPoints(Long.parseLong(data.get("prempoints")));
         ai.setUsedSpace(Long.parseLong(data.get("curspace")));
-        ai.setTrafficShareLeft((Long.parseLong(data.get("bodkb")) / 1000) * 1024l * 1024l);
+        ai.setTrafficShareLeft((long)(Long.parseLong(data.get("bodkb"))/1000.0) * 1024l*1024l);
 
         ai.setValidUntil(Long.parseLong(data.get("validuntil")) * 1000);
